pyproject.toml
[tool.poetry]
name = "idun-agent-manager"
version = "0.1.0"
description = "Monorepo for Idun Agent Platform"
authors = ["Geoffrey HARRAZI <geoffreyharrazi@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = []
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
idun-agent-engine = { path = "libs/idun_agent_engine", develop = true }
sqlalchemy = ">=2.0.36,<3.0.0"
psycopg = {version = ">=3.2.3,<4.0.0", extras = ["binary"]}
docker = ">=7.1.0,<8.0.0"
GitPython = ">=3.1.43,<4.0.0"
pydantic-settings = ">=2.6.1,<3.0.0"
python-multipart = ">=0.0.17,<0.1.0"
PyYAML = ">=6.0.2,<7.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.5"
black = "^24.8.0"
pre-commit = "^3.8.0"
mypy = "^1.10.0"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
types-PyYAML = "^6.0.12.20240808"
pyright = "^1.1.404"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
